fix: add steps.ts and add docstrings
