+++ src/client/workflow/auto-executor.ts
@@ -1,13 +1,7 @@
 import { QstashWorkflowAbort, QstashWorkflowError } from "../error";
 import type { WorkflowContext } from "./context";
-import type { ParallelCallState, Step } from "./types";
-import {
-  LazyFunctionStep,
-  LazySleepStep,
-  LazySleepUntilStep,
-  type AsyncStepFunction,
-  type BaseLazyStep,
-} from "./types";
+import type { AsyncStepFunction, ParallelCallState, Step } from "./types";
+import { LazyFunctionStep, LazySleepStep, LazySleepUntilStep, type BaseLazyStep } from "./steps";
 
 export class AutoExecutor {
   private context: WorkflowContext;
@@ -22,12 +16,35 @@ export class AutoExecutor {
     this.context = context;
   }
 
+  /**
+   * Adds a sleep step
+   *
+   * @param stepName
+   * @param sleep duration to sleep in seconds
+   * @returns
+   */
   public async addSleepStep(stepName: string, sleep: number) {
     return await this.addStep(new LazySleepStep(stepName, sleep));
   }
+
+  /**
+   * Adds a sleepUntil step
+   *
+   * @param stepName
+   * @param sleepUntil unix timestamp to wait until (in seconds)
+   * @returns
+   */
   public async addSleepUntilStep(stepName: string, sleepUntil: number) {
     return await this.addStep(new LazySleepUntilStep(stepName, sleepUntil));
   }
+
+  /**
+   * Adds an execution step
+   *
+   * @param stepName
+   * @param stepFunction function to execute
+   * @returns
+   */
   public async addFunctionStep<TResult>(
     stepName: string,
     stepFunction: AsyncStepFunction<TResult>
@@ -45,10 +62,10 @@ export class AutoExecutor {
    * If there is a single function, it's executed by itself. If there
    * are multiple, they are run in parallel.
    *
-   * @param asyncStepFunction step function to run in parallel
+   * @param stepInfo step plan to add
    * @returns result of the step function
    */
-  private async addStep<TResult>(StepInfo: BaseLazyStep<TResult>) {
+  private async addStep<TResult>(stepInfo: BaseLazyStep<TResult>) {
     this.stepCount += 1;
 
     const lazyStepList = this.activeLazyStepList ?? [];
@@ -58,7 +75,7 @@ export class AutoExecutor {
       this.indexInCurrentList = 0;
     }
 
-    lazyStepList.push(StepInfo);
+    lazyStepList.push(stepInfo);
     const index = this.indexInCurrentList++;
 
     const requestComplete = this.deferExecution().then(async () => {
@@ -83,9 +100,8 @@ export class AutoExecutor {
    * - If the step result is available in the steps, returns the result
    * - If the result is not avaiable, runs the function
    * - Sends the result to QStash
-   * - skips rest of the steps in the workflow in the current call
    *
-   * @param step runs a step by itself
+   * @param lazyStep lazy step to execute
    * @returns step result
    */
   private async runSingle<TResult>(lazyStep: BaseLazyStep<TResult>) {
@@ -109,11 +125,16 @@ export class AutoExecutor {
   private async runParallel<TResults extends unknown[]>(parallelSteps: {
     [K in keyof TResults]: BaseLazyStep<TResults[K]>;
   }): Promise<TResults> {
+    // get the step count before the parallel steps were added + 1
     const initialStepCount = this.stepCount - (parallelSteps.length - 1);
     const parallelCallState = this.getParallelCallState(parallelSteps.length, initialStepCount);
 
     switch (parallelCallState) {
       case "first": {
+        /**
+         * Encountering a parallel step for the first time, create plan steps for each parallel step
+         * and send them to QStash. QStash will call us back parallelSteps.length many times
+         */
         const planSteps = parallelSteps.map((parallelStep, index) =>
           parallelStep.getPlanStep(parallelSteps.length, initialStepCount + index)
         );
@@ -121,6 +142,12 @@ export class AutoExecutor {
         break;
       }
       case "partial": {
+        /**
+         * Being called by QStash to run one of the parallel steps. Last step in the steps list
+         * indicates which step is to be run
+         *
+         * Execute the step and call qstash with the result
+         */
         const planStep = this.context.steps.at(-1);
         if (!planStep || planStep.targetStep === 0) {
           throw new QstashWorkflowError(
@@ -133,9 +160,20 @@ export class AutoExecutor {
         break;
       }
       case "discard": {
+        /**
+         * We are still executing a parallel step but the last step is not a plan step, which means the parallel
+         * execution is in progress (other parallel steps are still running) but one of the parallel steps has
+         * called QStash with its result.
+         *
+         * This call to the API should be discarded: no operations are to be made. Parallel steps which are still
+         * running will finish and call QStash eventually.
+         */
         throw new QstashWorkflowAbort("discarded parallel");
       }
       case "last": {
+        /**
+         * All steps of the parallel execution have finished.
+         */
         // eslint-disable-next-line @typescript-eslint/no-magic-numbers
         const sortedSteps = this.context.steps.toSorted(
           (step, stepOther) => step.stepId - stepOther.stepId
@@ -183,15 +221,35 @@ export class AutoExecutor {
       return "first";
       // eslint-disable-next-line @typescript-eslint/no-magic-numbers
     } else if (remainingSteps.length >= 2 * parallelStepCount) {
+      // multipying by two since each step in parallel step will result in two
+      // steps: one plan step and one result step. If there are 3 parallel steps
+      // and steps list has at least 3*2=6 steps, the parallel steps have finished
       return "last";
     } else if (remainingSteps.at(-1)?.targetStep) {
+      // if the last step is a plan step, it means the step corresponding to the
+      // plan step will execute
       return "partial";
     } else {
+      // if the call is not the first/last/partial, it means that it's the result
+      // of one of the parallel calls but others are still running, meaning that
+      // the current call should be discarded
       return "discard";
     }
   }
 
+  /**
+   * sends the steps to QStash as batch
+   *
+   * @param steps steps to send
+   */
   private async submitStepsToQstash(steps: Step[]) {
+    // if there are no steps, something went wrong. Raise exception
+    if (steps.length === 0) {
+      throw new QstashWorkflowError(
+        `Unable to submit steps to Qstash. Provided list is empty. Current step: ${this.stepCount}`
+      );
+    }
+
     await this.context.client.batchJSON(
       steps.map((singleStep) => {
         return {
@@ -205,21 +263,29 @@ export class AutoExecutor {
       })
     );
 
-    const error =
-      steps.length > 0
-        ? new QstashWorkflowAbort(steps[0].stepName, steps[0])
-        : new QstashWorkflowError(
-            `Unable to submit steps to Qstash. Provided list is empty. Current step: ${this.stepCount}`
-          );
-    throw error;
+    // if the steps are sent successfully, abort to stop the current request
+    throw new QstashWorkflowAbort(steps[0].stepName, steps[0]);
   }
 
+  /**
+   * Get the promise by executing the lazt steps list. If there is a single
+   * step, we call `runSingle`. Otherwise `runParallel` is called.
+   *
+   * @param lazyStepList steps list to execute
+   * @returns promise corresponding to the execution
+   */
   private getExecutionPromise(lazyStepList: BaseLazyStep[]): Promise<unknown> {
     return lazyStepList.length === 1
       ? this.runSingle(lazyStepList[0])
       : this.runParallel(lazyStepList);
   }
 
+  /**
+   * @param lazyStepList steps we executed
+   * @param result result of the promise from `getExecutionPromise`
+   * @param index index of the current step
+   * @returns result[index] if lazyStepList > 1, otherwise result
+   */
   private static getResult<TResult>(lazyStepList: BaseLazyStep[], result: unknown, index: number) {
     if (lazyStepList.length === 1) {
       return result as TResult;
