+++ src/client/workflow/types.ts
@@ -36,6 +36,9 @@ export type WorkflowServeParameters<TInitialPayload, TResponse extends Response 
   options?: WorkflowServeOptions<TResponse, TInitialPayload>;
 };
 
+/**
+ * Function parsing initial payload from string to an object
+ */
 export type InitialPayloadParser<TInitialPayload = unknown> = (
   initialPayload: string
 ) => TInitialPayload;
@@ -44,85 +47,19 @@ export type WorkflowServeOptions<
   TResponse extends Response = Response,
   TInitialPayload = unknown,
 > = {
+  /**
+   * QStash client
+   */
   client?: Client;
+  /**
+   * Function called to return a response after each step execution
+   *
+   * @param workflowId
+   * @returns response
+   */
   onFinish?: (workflowId: string) => TResponse;
+  /**
+   * Function to parse the initial payload passed by the user
+   */
   initialPayloadParser?: InitialPayloadParser<TInitialPayload>;
 };
-
-/**
- *
- */
-export abstract class BaseLazyStep<TResult = unknown> {
-  public readonly stepName;
-  constructor(stepName: string) {
-    this.stepName = stepName;
-  }
-
-  public getPlanStep(concurrent: number, targetStep: number): Step<undefined> {
-    return {
-      stepId: 0,
-      stepName: this.stepName,
-      concurrent,
-      targetStep,
-    };
-  }
-
-  public abstract getResultStep(stepId: number): Promise<Step<TResult>>;
-}
-
-export class LazyFunctionStep<TResult = unknown> extends BaseLazyStep<TResult> {
-  private readonly stepFunction: AsyncStepFunction<TResult>;
-
-  constructor(stepName: string, stepFunction: AsyncStepFunction<TResult>) {
-    super(stepName);
-    this.stepFunction = stepFunction;
-  }
-
-  public async getResultStep(stepId: number): Promise<Step<TResult>> {
-    const result = await this.stepFunction();
-
-    return {
-      stepId,
-      stepName: this.stepName,
-      out: result,
-      concurrent: 1,
-      targetStep: 0,
-    };
-  }
-}
-
-export class LazySleepStep extends BaseLazyStep {
-  private readonly sleep: number;
-
-  constructor(stepName: string, sleep: number) {
-    super(stepName);
-    this.sleep = sleep;
-  }
-  public async getResultStep(stepId: number): Promise<Step> {
-    return await Promise.resolve({
-      stepId,
-      stepName: this.stepName,
-      sleepFor: this.sleep,
-      concurrent: 1,
-      targetStep: 0,
-    });
-  }
-}
-
-export class LazySleepUntilStep extends BaseLazyStep {
-  private readonly sleepUntil: number;
-
-  constructor(stepName: string, sleepUntil: number) {
-    super(stepName);
-    this.sleepUntil = sleepUntil;
-  }
-  public async getResultStep(stepId: number): Promise<Step> {
-    return await Promise.resolve({
-      stepId,
-      stepName: this.stepName,
-      sleepUntil: this.sleepUntil,
-      concurrent: 1,
-      targetStep: 0,
-    });
-  }
-}
